requirements.txt
langchain==0.1.9
langchain-google-genai==0.0.6
google-generativeai==0.3.2
pandas==2.2.0
python-dotenv==1.0.1
streamlit==1.31.0
